refactor: refactor the tui structure
